requirements.txt
autopep8==1.6.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
futures==3.0.5
Jinja2==3.1.4
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.5.5.62
packaging==24.0
Pillow==8.4.0
pycodestyle==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0

six==1.16.0
sympy==1.12.1
toml==0.10.2
torch==2.3.1
typing_extensions==4.12.1
